tg_bot/pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.7",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.9.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "isort>=5.13.0",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
